Only do it when > 100 execs

+++ dkron/store.go
@@ -225,10 +225,12 @@ func (s *Store) SetExecution(execution *Execution) (string, error) {
 	sort.Sort(byGroup)
 
 	// Delete all execution results over the limit, starting from olders
-	for i := range byGroup[99:] {
-		err := s.Client.Delete(fmt.Sprintf("%s/executions/%s/%s", s.keyspace, execs[i].JobName, execs[i].Key()))
-		if err != nil {
-			log.Errorf("store: Trying to delete overflowed execution %s", execs[i].Key())
+	if len(byGroup) > 99 {
+		for i := range byGroup[99:] {
+			err := s.Client.Delete(fmt.Sprintf("%s/executions/%s/%s", s.keyspace, execs[i].JobName, execs[i].Key()))
+			if err != nil {
+				log.Errorf("store: Trying to delete overflowed execution %s", execs[i].Key())
+			}
 		}
 	}
